Add locks.cpp
