+++ src/scheduler/Executor.cpp
@@ -302,7 +302,7 @@ void Executor::threadPoolThread(int threadPoolIdx)
         // In all threads if we've finished with a ptp group, we need to tidy up
         // to ensure all sockets can be destructed
         if (msg.groupid() > 0) {
-            broker.resetThreadLocalCache(true);
+            broker.resetThreadLocalCache(false);
 
             // Remove the group from this host if last in batch
             if (isLastInBatch) {
