+++ src/util/CMakeLists.txt
@@ -12,6 +12,7 @@ faabric_lib(util
     gids.cpp
     json.cpp
     latch.cpp
+    locks.cpp
     logging.cpp
     memory.cpp
     network.cpp
